requirements.txt
python-dotenv==1.2.1
aiogram==3.23.0
aiohttp==3.13.2
requests==2.32.5
aiosqlite==0.22.1
SQLAlchemy==2.0.45
yookassa==3.9.0